compute_vocab: Reset order of reading after filtering on min_freq
In the previous implementation of compute_vocab, when min_freq was set but not max_freq (most of time this is the case only for the tasks sets of labels), the function was using reading order (=order of dictionary self.freqs as it is built by reading the training data line by line) instead of order of frequency. The order of frequency was already used, but only when min_freq was not set (i.e. max_size or nothing)

In order to preserve reproducibility of older experiments, this commit resets this reading order when min_freq is set, so that the steps to compute the vocabulary (or list of labels) are:
- compute frequencies by reading the training data line by line
- sort candidate entries by frequency
- (if max_size is set) keep only the n most frequent entries
- (if min_freq is set) filter out entries with a low frequency
- (if min_freq is set) change order of entries to reflect order of freqs dict