Change default settings for min_freq
In order to preserve reproducibility with older experiments,
we need to compute vocab in reading order for tasks and
frequency order for word/char (unless only min_freq is set).

This commit sets the default values in the JSON default_settings.json to false for word_min_freq/char_min_freq to *false*, and min_freq for all tasks to 1. In parallel, the default value of the min_freq parameter when initializing a LabelEncoder object is None instead of 1.