Adapt TaskScheduler to pretraining task
When no decoder target task is set (i.e. pretraining mode), the TaskScheduler can now be used by setting the lm_fwd task as target and ignoring the task weighting updates.

Also, when the training finishes, the best model is now loaded also in the case of pretraining (since there is a TaskScheduler again)

+ show early stopping epoch number in the early stopping message