Minor variable name change when updating the lr_scheduler
When a target task is defined, the evaluation score (by default based on accuracy) is used (expected mode max), whereas in pretraining mode the sum of lm losses are used (expected mode min).
Therefore lr_scheduler_score is a more appropriate variable name as lr_scheduler_loss because it should be the most frequent usage of this variable.