Make LM decoder loss ignore UNK targets
In order to avoid making the LM learn to predict UNK, ignore word targets that are unknown in the training data when computing the loss (on top of the PAD token already ignored).

Use value -100 in the lm_fwd_decoder and lm_bwd_decoder for the ignore_index loss param (yet keep the default value for other decoder, i.e. index of the PAD token)